+++ resources/views/components/button-icon.blade.php
@@ -8,7 +8,7 @@
 
     $wireTarget = $attributes->get('wire:target') ?? null;
 
-    if($spinner && is_null($wireTarget)) {
+    if ($spinner && is_null($wireTarget)) {
         $wireTarget = $attributes->get('wire:click') ?? null;
     }
 
@@ -17,37 +17,46 @@
 @endphp
 
 <div class="relative">
-  <{{ $tag }}
-    {{ $attributes->merge([
-      'class' => 'flex items-center justify-center rounded-'.$rounded.' shadow-lg transition '.$colorPreset['bg'] . ' ' . $colorPreset['text'] . ' ' . $colorPreset['bg_hover'] . ' ' . $sizePreset['buttonIcon']
-    ]) }}
-    @if($spinner && $wireTarget)
-        wire:target="{{ $wireTarget }}"
-    @endif
-    @if ($isLink)
-      href="{{ $attributes->get('href') }}"
-    @endif
+    <{{ $tag }}
+        {{ $attributes->merge([
+            'class' =>
+                'flex items-center justify-center cursor-pointer rounded-' .
+                $rounded .
+                ' shadow-lg transition ' .
+                $colorPreset['bg'] .
+                ' ' .
+                $colorPreset['text'] .
+                ' ' .
+                $colorPreset['bg_hover'] .
+                ' ' .
+                $sizePreset['buttonIcon'],
+        ]) }}
+        @if ($spinner && $wireTarget) wire:target="{{ $wireTarget }}" @endif
+        @if ($isLink) href="{{ $attributes->get('href') }}" @endif
+        {{ $attributes->except('href', 'class') }}>
 
-    {{ $attributes->except('href', 'class') }}
-  >
-
-    @if($spinner && $wireTarget)
-        <div wire:loading wire:target="{{ $wireTarget }}">
-            @include('beartropy-ui-svg::beartropy-spinner', ['class' => 'w-5 h-5 animate-spin', 'tabindex' => '-1'])
-        </div>
-        <div wire:loading.remove wire:target="{{ $wireTarget }}">
-            @if ($slot->isNotEmpty())
-            {!! $slot !!}
-            @else
-                <x-beartropy-ui::icon name="{{$icon}}" class="{{$sizePreset['buttonIconIcon']}}" set="{{$iconSet}}" variant="{{$iconVariant}}" />
-            @endif
-        </div>
-    @else
-        @if ($slot->isNotEmpty())
-            {!! $slot !!}
+        @if ($spinner && $wireTarget)
+            <div wire:loading wire:target="{{ $wireTarget }}">
+                @include('beartropy-ui-svg::beartropy-spinner', [
+                    'class' => 'w-5 h-5 animate-spin',
+                    'tabindex' => '-1',
+                ])
+            </div>
+            <div wire:loading.remove wire:target="{{ $wireTarget }}">
+                @if ($slot->isNotEmpty())
+                    {!! $slot !!}
+                @else
+                    <x-beartropy-ui::icon name="{{ $icon }}" class="{{ $sizePreset['buttonIconIcon'] }}"
+                        set="{{ $iconSet }}" variant="{{ $iconVariant }}" />
+                @endif
+            </div>
         @else
-            <x-beartropy-ui::icon name="{{$icon}}" class="{{$sizePreset['buttonIconIcon']}}" set="{{$iconSet}}" variant="{{$iconVariant}}" />
+            @if ($slot->isNotEmpty())
+                {!! $slot !!}
+            @else
+                <x-beartropy-ui::icon name="{{ $icon }}" class="{{ $sizePreset['buttonIconIcon'] }}"
+                    set="{{ $iconSet }}" variant="{{ $iconVariant }}" />
+            @endif
         @endif
-    @endif
-  </{{ $tag }}>
+        </{{ $tag }}>
 </div>